vmlinux: introduce vmlinux-flavored.h to contain flavored types
